requirements.txt
django
re
psycopg2
requests
bs4
djangorestframework
django-rest-auth
